refactor: 1,2단계 미션 컴포넌트 구분
